+++ zrml/prediction-markets/src/old_tests.rs
@@ -142,104 +142,6 @@ fn simple_create_scalar_market(
 }
 
 #[test]
-fn admin_move_market_to_resolved_resolves_reported_market() {
-    // NOTE: Bonds are always in ZTG, irrespective of base_asset.
-    let test = |base_asset: Asset<MarketId>| {
-        let end = 33;
-        simple_create_categorical_market(
-            base_asset,
-            MarketCreation::Permissionless,
-            0..end,
-            ScoringRule::Lmsr,
-        );
-        let market_id = 0;
-
-        // Give ALICE `SENTINEL_AMOUNT` free and reserved ZTG; we record the free balance to check
-        // that the correct bonds are unreserved!
-        assert_ok!(AssetManager::deposit(Asset::Ztg, &ALICE, 2 * SENTINEL_AMOUNT));
-        assert_ok!(Balances::reserve_named(
-            &PredictionMarkets::reserve_id(),
-            &ALICE,
-            SENTINEL_AMOUNT
-        ));
-        let balance_free_before = Balances::free_balance(ALICE);
-        let balance_reserved_before =
-            Balances::reserved_balance_named(&PredictionMarkets::reserve_id(), &ALICE);
-
-        let market = MarketCommons::market(&0).unwrap();
-        let grace_period = end + market.deadlines.grace_period;
-        run_to_block(grace_period + 1);
-        let category = 1;
-        let outcome_report = OutcomeReport::Categorical(category);
-        assert_ok!(PredictionMarkets::report(
-            RuntimeOrigin::signed(BOB),
-            market_id,
-            outcome_report.clone()
-        ));
-        assert_ok!(PredictionMarkets::admin_move_market_to_resolved(
-            RuntimeOrigin::signed(SUDO),
-            market_id
-        ));
-
-        let market = MarketCommons::market(&market_id).unwrap();
-        assert_eq!(market.status, MarketStatus::Resolved);
-        assert_eq!(market.report.unwrap().outcome, outcome_report);
-        assert_eq!(market.resolved_outcome.unwrap(), outcome_report);
-        System::assert_last_event(
-            Event::MarketResolved(market_id, MarketStatus::Resolved, outcome_report).into(),
-        );
-
-        assert_eq!(
-            Balances::reserved_balance_named(&PredictionMarkets::reserve_id(), &ALICE),
-            balance_reserved_before
-                - <Runtime as Config>::OracleBond::get()
-                - <Runtime as Config>::ValidityBond::get()
-        );
-        assert_eq!(
-            Balances::free_balance(ALICE),
-            balance_free_before
-                + <Runtime as Config>::OracleBond::get()
-                + <Runtime as Config>::ValidityBond::get()
-        );
-    };
-    ExtBuilder::default().build().execute_with(|| {
-        test(Asset::Ztg);
-    });
-    #[cfg(feature = "parachain")]
-    ExtBuilder::default().build().execute_with(|| {
-        test(Asset::ForeignAsset(100));
-    });
-}
-
-#[test_case(MarketStatus::Active)]
-#[test_case(MarketStatus::Closed)]
-#[test_case(MarketStatus::Resolved)]
-fn admin_move_market_to_resolved_fails_if_market_is_not_reported_or_disputed(
-    market_status: MarketStatus,
-) {
-    ExtBuilder::default().build().execute_with(|| {
-        simple_create_categorical_market(
-            Asset::Ztg,
-            MarketCreation::Permissionless,
-            0..33,
-            ScoringRule::Lmsr,
-        );
-        let market_id = 0;
-        assert_ok!(MarketCommons::mutate_market(&market_id, |market| {
-            market.status = market_status;
-            Ok(())
-        }));
-        assert_noop!(
-            PredictionMarkets::admin_move_market_to_resolved(
-                RuntimeOrigin::signed(SUDO),
-                market_id,
-            ),
-            Error::<Runtime>::InvalidMarketStatus,
-        );
-    });
-}
-
-#[test]
 fn it_creates_binary_markets() {
     ExtBuilder::default().build().execute_with(|| {
         simple_create_categorical_market(
